avoid warning

+++ src/NVector.cc
@@ -106,7 +106,7 @@ NVector operator*(NVector op, const double scale)
 
 std::ostream& operator<<(std::ostream& out, const NVector& op)
 {
-  for ( int i=0; i<op.nDim-1; ++i ) out << op.p(i) << ", ";
+  for ( size_t i=0; i<op.nDim-1; ++i ) out << op.p(i) << ", ";
   out << op.p(op.nDim-1);
 
   return out;
